fixed go fail

+++ toFloat.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 )
 
-func ToFloat()(float, error){
+func ToFloat()(float64, error){
 	reader := bufio.NewReader(os.Stdin)
 	input, err := reader.ReadString('\n')
 	if err!=nil{
